+++ packages/node/CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Filter out transactions that can't be decoded (#280)
 
+### Change
+- Support bigint filter
+
 ## [4.1.0] - 2024-08-12
 ### Added
 - Support for endpoint configs (#278)